persist: fix bug where we'd incorrectly delete an initial rollup
This was causing maelstrom to flake in CI.

If two shard users race to initialize it, one will write a rollup and
only find out that it lost when attempting to CaS it in to the initial
consensus write. To avoid leaking this blob, we delete it when that
happens.

However, if we get an Indeterminate error in the CaS, but it actually
went through, then we'll "contend" with ourselves and get an expectation
mismatch. Previously, this would lead to us incorrectly deleting the
rollup blob, which manifests as a permanently hung shard. Instead, use
the actual fetched state to determine if our rollup actually made it in
and decide whether to delete based on that.

Closes #14575

+++ src/persist-client/src/internal/state_versions.rs
@@ -146,11 +146,27 @@ impl StateVersions {
                 // We lost a CaS race and someone else initialized the shard,
                 // use the value included in the CaS expectation error.
 
-                // Clean up the rollup blob that we were trying to reference.
-                let (_, rollup_key) = initial_state.latest_rollup();
-                self.delete_rollup(&shard_id, rollup_key).await;
+                let state = self.fetch_current_state(&shard_id, live_diffs).await;
 
-                return self.fetch_current_state(&shard_id, live_diffs).await;
+                // Clean up the rollup blob that we were trying to reference.
+                //
+                // SUBTLE: If we got an Indeterminate error in the CaS above,
+                // but it actually went through, then we'll "contend" with
+                // ourselves and get an expectation mismatch. Use the actual
+                // fetched state to determine if our rollup actually made it in
+                // and decide whether to delete based on that.
+                let (_, rollup_key) = initial_state.latest_rollup();
+                let should_delete_rollup = match state.as_ref() {
+                    Ok(state) => !state.collections.rollups.values().any(|x| x == rollup_key),
+                    // If the codecs don't match, then we definitely didn't
+                    // write the state.
+                    Err(CodecMismatch { .. }) => true,
+                };
+                if should_delete_rollup {
+                    self.delete_rollup(&shard_id, rollup_key).await;
+                }
+
+                return state;
             }
         }
     }
